requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
future==0.18.3
httplib2==0.22.0
idna==3.4
gqlalchemy==1.4.1
matplotlib==3.7.2
nebula3-python==3.4.0
#neo4j==5.7.0
numpy==1.24.2
pandas==2.0.0
plyer==2.1.0
pyobjus==1.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
